+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
-## [2.0.0] 2023-10-27
+## [1.2.0] 2023-10-27
 - Fixed Button Spacing
 
 ## [1.1.9] 2023-10-18